apps/chatbot_slack/requirements.txt
slack_bolt
confluent_kafka
python-dotenv